refactor: implement custom errors in errors.rs
